my_learning/todo_with_poetry_dockerize/backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Aqeel Shahzad <67902032+aqeel-spec@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.110.2"
sqlmodel = "^0.0.16"
uvicorn = {extras = ["standard"], version = "^0.29.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-multipart = "^0.0.9"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
pytest = "^8.1.1"
httpx = "^0.27.0"
psycopg = {extras = ["binary"], version = "^3.1.18"}


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
